Refactor: Organize code into modules and directories
This commit restructures the project by breaking down the monolithic JavaScript and CSS files into smaller, more manageable modules. This improves organization, maintainability, and adheres to modern JavaScript module patterns.
