+++ css/components.css
@@ -1,0 +1,118 @@
+/* Small reusable components */
+
+/* Icons */
+.icon {
+  width: 20px;
+  height: 20px;
+  flex: 0 0 auto;
+}
+
+/* Buttons */
+.btn {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.55rem;
+  border-radius: 999px;
+  padding: 0.9rem 1.15rem;
+  border: 1px solid transparent;
+  background: transparent;
+  color: rgba(255, 255, 255, 0.90);
+  text-decoration: none;
+  cursor: pointer;
+  transition: transform var(--transition), background var(--transition), border-color var(--transition), color var(--transition), box-shadow var(--transition);
+  user-select: none;
+  -webkit-tap-highlight-color: transparent;
+}
+
+.btn span {
+  letter-spacing: -0.01em;
+}
+
+.btn:focus-visible {
+  outline: 2px solid rgba(34, 211, 238, 0.75);
+  outline-offset: 3px;
+}
+
+.btn-sm {
+  padding: 0.65rem 0.9rem;
+  font-size: 0.95rem;
+}
+
+.btn-lg {
+  padding: 0.95rem 1.2rem;
+  font-size: 1rem;
+}
+
+.btn-primary {
+  background: linear-gradient(90deg, rgba(124, 58, 237, 0.95), rgba(34, 211, 238, 0.92));
+  border-color: rgba(255, 255, 255, 0.08);
+  box-shadow: 0 18px 60px rgba(124, 58, 237, 0.25);
+  color: rgba(7, 8, 11, 0.92);
+}
+
+.btn-primary:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 22px 76px rgba(34, 211, 238, 0.20);
+}
+
+.btn-secondary,
+.btn-outline {
+  border-color: rgba(255, 255, 255, 0.16);
+  background: rgba(255, 255, 255, 0.03);
+}
+
+.btn-secondary:hover,
+.btn-outline:hover {
+  transform: translateY(-2px);
+  border-color: rgba(255, 255, 255, 0.26);
+  background: rgba(255, 255, 255, 0.06);
+}
+
+/* Section head */
+.section-head {
+  display: grid;
+  gap: 0.75rem;
+  margin-bottom: clamp(1.8rem, 3.4vw, 3rem);
+}
+
+.section-title {
+  font-family: var(--font-display);
+  font-size: var(--h2);
+  line-height: 1.1;
+  letter-spacing: -0.035em;
+  margin: 0;
+}
+
+.section-lede {
+  margin: 0;
+  max-inline-size: 70ch;
+  font-size: var(--lede);
+  color: rgba(255, 255, 255, 0.70);
+  text-wrap: pretty;
+}
+
+/* Cursor glow (desktop only via JS) */
+.cursor-glow {
+  position: fixed;
+  left: 0;
+  top: 0;
+  width: 560px;
+  height: 560px;
+  border-radius: 999px;
+  pointer-events: none;
+  z-index: 6;
+  opacity: 0;
+  transform: translate3d(-1000px, -1000px, 0);
+  background:
+    radial-gradient(circle at 30% 30%, rgba(34, 211, 238, 0.16), transparent 60%),
+    radial-gradient(circle at 65% 40%, rgba(124, 58, 237, 0.14), transparent 58%);
+  filter: blur(26px);
+  mix-blend-mode: screen;
+  transition: opacity 240ms var(--ease);
+}
+
+.js .cursor-glow {
+  opacity: 1;
+}
+
