+++ index.html
@@ -7,7 +7,7 @@
     <title>Nenvy — Design is a language for problem solving.</title>
     <meta name="description" content="Nenvy is a design-first web studio building modern digital systems—focused on clarity, automation, and brand." />
     <link rel="icon" type="image/png" href="icon-black.png">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="css/main.css">
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300..900&family=Inter+Tight:wght@300..900&display=swap" rel="stylesheet">
@@ -442,7 +442,7 @@
     <script src="https://cdn.jsdelivr.net/npm/gsap@3.13.0/dist/gsap.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/gsap@3.13.0/dist/ScrollTrigger.min.js"></script>
     <script src="https://www.paypal.com/sdk/js?client-id=ATnObauWTeaii36d5zifRVm7VcG5a34SO1BW0vBdMA7vTOD7hGOoPPkcJEb-RBsSuuVI27oWVSTa2SJ5&vault=true&intent=subscription" data-sdk-integration-source="button-factory"></script>
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
     <script>
       paypal.Buttons({
           style: {
