+++ js/features/smooth-scroll.js
@@ -1,0 +1,18 @@
+import { qsa } from '../utils/dom.js';
+
+export function initSmoothScroll({ reducedMotion }) {
+  qsa('a[href^="#"]').forEach((anchor) => {
+    anchor.addEventListener('click', (e) => {
+      const href = anchor.getAttribute('href');
+      if (!href || href === '#') return;
+      const target = document.querySelector(href);
+      if (!target) return;
+
+      e.preventDefault();
+      const behavior = reducedMotion ? 'auto' : 'smooth';
+      target.scrollIntoView({ behavior, block: 'start' });
+      history.pushState(null, '', href);
+    });
+  });
+}
+
