+++ js/utils/dom.js
@@ -1,0 +1,8 @@
+export function qs(sel, root = document) {
+  return root.querySelector(sel);
+}
+
+export function qsa(sel, root = document) {
+  return Array.from(root.querySelectorAll(sel));
+}
+
